demo_website_v2/backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
newsapi-python>=0.2.7,<1.0
youtube-transcript-api>=0.6,<1.0
youtube-search-python>=1.6,<2.0
pydantic>=2.6,<3.0
httpx>=0.24.1,<0.25
